requirements.txt
black==21.9b0
celery==5.1.2
dj-database-url==0.5.0
Django==3.2.8
djangorestframework==3.12.4
flake8==4.0.1
psycopg2-binary==2.9.3
drf-nested-routers==0.93.4
gunicorn
dj-database-url==0.5.0
whitenoise== 6.0.0